+++ package.json
@@ -39,6 +39,7 @@
     "leaflet-color-number-markers": "^1.0.7",
     "leaflet-geosearch": "^3.6.0",
     "lint-staged": "^11.2.6",
+    "lodash.isequal": "^4.5.0",
     "nanoid": "^3.3.1",
     "nodemailer": "^6.7.2",
     "papaparse": "^5.3.1",
@@ -56,6 +57,7 @@
     "react-leaflet": "^3.1.0",
     "react-leaflet-markercluster": "^3.0.0-rc1",
     "react-papaparse": "^4.0.2",
+    "react-reflex": "^4.0.9",
     "react-router": "5.2.0",
     "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
